Do some refactoring, get into working order, and add inputs for settings to page.
